[REFACTOR] UserService 인터페이스화를 위한 패키지 이동
